added effectiveGasPrice
